Preparing for having a test database
We want to differentiate between production models and models
for test cases. We can do that by giving a particular service a
higher rank. However, if we have multiple cardinality, only influence
the order in which the services are loaded. We cannot choose a particular
model provider, hence removing this annotation.