Second test
Another problem emerged: since the migratorservice selects the migrator
based on sourceversion (and nothing else), we cannot control directly
which migrator is used. This is unfortunate if we want to test different
migrators.
Hence, as a workaround, we implement ONE migrator, tell the migrator
through properties which test case is performing the migration,
and then let the migrator decide, based on testcase, which migration to perform.