Delay the retrieval of model packages
This allows us to replace the model as needed, as the model is not
retrieved when the service is started, but when they are actually used,
i.e. in the startup phase of the service. This allows us to shutdown the
service, replace the model, and start it up again.