Checking authorization
Access to a resource is controlled by checking whether the project
component of the URI matches the project the user has been authenticated
for.

This is a problem for integration tests. They bootstrap resources, so no
request will be authorized since there is nothing to check against.
Hence, we provide an authenticate method for test use only.