API change in Rest Service
Added an additional parameter, token, to all REST operations as this allows us
to provide fine-grained access control to resources through the session service.

The token is extracted from the request header. If we need to control access to
a resource, we can simply inject the session service, retrieve access rights and
allow/deny access correspondingly.