Reverting again to create store objects
It turns out we cannot reuse the store. If persistency is restarted, we would
reuse an invalid store object (CDO closes it). In this case, looking at the
client code, I think it's best to let CDO decide when to release the store and
we just create new ones. I renamed the method to make this more clear.