Unintentional multiple changes
If a commit fails, the change, executed in doChange() is executed multiple
times. The caller of doAndCommit() would have to make sure that the change
operations are executed only once. I don't think this was the actually intended
behavior of semantics of doAndCommit(). This modification executes doChange()
only once, independent of when the commit succeeds.