Do and commit (#268)
* Replace commit with doAndCommit calls

* Remove commit() from the API to prevent use

* Unintentional multiple changes

If a commit fails, the change, executed in doChange() is executed multiple
times. The caller of doAndCommit() would have to make sure that the change
operations are executed only once. I don't think this was the actually intended
behavior of semantics of doAndCommit(). This modification executes doChange()
only once, independent of when the commit succeeds.

* Revert "Unintentional multiple changes"

This commit was based upon the wrong assumption that model changes are *not*
rolled back after a transaction commit failure.

This reverts commit f867899cb22a4e86857854606a42e41f9027c23a.

* Move model changing operations into doChange()

These operations are rolled back in case the commit does not succeed.

* Fix test