Remove useless variable 'compoundId'
This is a code smell.
The variable 'compoundId' is announced as 'let compoundId: string = lastCommand.compoundId;', but in this function 'lastCommand.compoundId' is only used once. We think there is no need to replace 'lastCommand.compoundId' with 'compoundId' as it only used once in the function. So we decide to remove 'compoundId' as a useless variable.

Group 2