Refactor this function so that its implementation doesn't duplicate t… (#279)
* Refactor this function so that its implementation doesn`t duplicate the one on line 83

Because InputParameters and OutputParameters have same  inplementation, so I create getInOutParameters to return outParameters to call. The functions that we change are select and click functions in Test-case-row. We have manually test the functions and it can work as before.

* Refactor the inputParameters

I refactor the inputParameters because i forgot remove the duplicated code in method inputParameters().