Removed the duplication "confirmDelete" function on line 18
The functionalities "confirmdelete" are being called by another function "openOkcancel" which results in duplication. The code has been refactored by changing the "openOkcancel " recurrences to "confirmDelete". The functionalities of "openOcancel" are not being used anywhere but only in this code segment. Duplication has been removed as code clone incur the maintainability costs and effort.