Print view for models
We add "hide UI" button in card-header model. After clicking this button, all UI elements except model view are invisible.  And when clicking the button again, the UI will back to the original view. And we use manual test to test this button. We try to hide the tool palette, but we fail.