Update objects.ts
In the refactored code we have created conditions in such a way that if one the of the functions return false then it goes back to main function. If all three statement are true and return true then it start comparing the objects. Which increases the understandability of the code.